api/requirements.txt
tensorflow==2.21.0
tensorflow-serving-api==2.19.1
fastapi==0.136.0
uvicorn[standard]==0.44.0
python-multipart==0.0.20
pillow==11.2.1
matplotlib==3.10.1
numpy==2.2.5
